Extract ImapPushState from ImapStore
